requirements.txt
astroid==3.3.11
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
dill==0.4.0
executing==2.2.0
Faker==37.6.0
flake8==7.3.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
parso==0.8.5
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.8
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.0.2
six==1.17.0
stack-data==0.6.3
tomlkit==0.13.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
